Read configuration as tracked in previous run
This commit makes sure that the previous run's configuration path
is saved. It also makes sure the path is read at the previous run's
revision to match existing headers based on previous run's config
data.

It also makes sure that the tracking file is not touched when there
is no file to process.

Fixes #27.